去除一些警告

+++ MoblieCQUPT_iOS/MoblieCQUPT_iOS/Tool/NetWork/ORWRequestCache.m
@@ -95,7 +95,7 @@ static const NSInteger ORWDeafultCacheTime = 60*60*24;
     if (fectchDictory) {
         NSInteger cacheDeadTime = (NSInteger)[fectchDictory objectForKey:ORWDeadtimeCol];
         NSInteger nowTime = [[NSDate date] timeIntervalSince1970];
-        NSLog(@"比较:现在%d-记录%d",nowTime,cacheDeadTime);
+        NSLog(@"比较:现在%ld-记录%ld",nowTime,(long)cacheDeadTime);
         return (cacheDeadTime>nowTime?YES:NO);
     }
